test(activemodel): move TS-only decimal dirty case out of the Rails-mapped file
Keeps test:compare's "extra (TS only)" and "wrong describe" counts at their
baseline: the model-level revert case moves to decimal.trails.test.ts, and the
new date-time serialize_cast_value test moves into the DateTimeTest describe
rather than the trailing DateTimeType#isChanged one.
